move dash Gas component over to resources, delete dupes
