+++ app/App/Main/Account/style/index.styl
@@ -1719,7 +1719,6 @@
 
 
 @import '../Balances/style'
-@import '../Gas/style'
 @import '../Inventory/style'
 @import '../Permissions/style'
 @import '../Requests/style'
