+++ dash/App/Chains/style/index.styl
@@ -1107,6 +1107,5 @@
     z-index 1
     margin-top 0px
 
-@import '../Gas/style/index.styl'
 @import '../Usage/style/index.styl'
 @import '../Tokens/style/index.styl'
