+++ resources/Components/Gas/style/index.styl
@@ -1,192 +1,3 @@
-.sliceContainer
-  background var(--ghostZ)
-  position relative
-  margin 0px 6px 12px 6px
-  border-radius 12px
-  padding 3px 0px 3px 0px
-  font-size 14px
-  font-weight 300
-
-  .sliceContainerClose
-    cursor pointer
-    height 16px
-    display flex
-    justify-content center
-    align-items center
-    border-radius 8px
-    color black
-    div
-      pointer-events none
-      position relative
-      margin 0px 18px
-      top 2px
-
-  .sliceContainerClose:hover
-    color var(--white)
-
-  .sliceTileCenter
-    justify-content center
-
-  .sliceTile
-    cursor pointer
-    display flex
-    justify-content space-between
-    align-items center
-    transition var(--standardFast)
-    transform translateZ(0)
-    // margin 3px 3px 0px 3px
-    transform scale(1)
-    background var(--ghostA)
-    position relative
-    border-radius 9px
-    height 42px
-    box-sizing border-box
-    font-size 12px
-    font-weight 300
-    *
-      pointer-events none
-
-    .sliceTileConnectionName
-      padding 0px 39px
-      color var(--outerspace)
-      text-transform uppercase
-      font-size 13px
-      font-weight 400
-      letter-spacing 1px
-
-    .sliceTileBlock
-      display flex
-      align-items center
-      padding-right 11px
-      font-family FiraCode, monospace
-      font-size 14px
-      font-weight 600
-      position relative
-      top 1px
-      bottom 0px
-
-      .sliceTileBlockIcon
-        color var(--good)
-        padding-right 5px
-
-      .sliceTileChainId
-        padding-right 15px
-
-    .sliceTileIndicator
-      color var(--outerspace)
-      cursor pointer
-      position absolute 
-      top 50%
-      margin-top -2px
-      left 13px
-      bottom 0px
-      width 5px
-      height 5px
-      border-radius 3px
-      display flex
-      justify-content center
-      align-items center
-      border-radius 12px
-      background var(--good)
-      *
-        pointer-events none
-
-
-    .sliceTileIndicatorLarge
-      color var(--outerspace)
-      cursor pointer
-      position absolute 
-      top 50%
-      margin-top -5px
-      left 14px
-      bottom 0px
-      width 10px
-      height 10px
-      border-radius 4px
-      display flex
-      justify-content center
-      align-items center
-      border-radius 12px
-      background var(--good)
-      *
-        pointer-events none
-
-    .sliceTileIndicatorActive
-      background var(--goodHighlight)
-      box-shadow 0px 0px 4px var(--goodHighlight)
-
-    .sliceTileIndicatorGood
-      background var(--good)
-      box-shadow 0px 0px 4px var(--good)
-
-    .sliceTileIndicatorPending
-      background var(--pending)
-      box-shadow 0px 0px 4px var(--pending)
-
-    .sliceTileIndicatorBad
-      background var(--bad)
-      box-shadow 0px 0px 4px var(--bad)
-
-    .sliceTileChangeChain
-      color var(--outerspace)
-      cursor pointer
-      position absolute 
-      top 0px
-      right 0px
-      bottom 0px
-      width 41px
-      display flex
-      justify-content center
-      align-items center
-      border-radius 12px
-      *
-        pointer-events none
-
-    .sliceTileChangeChain:hover
-      background var(--ghostB)
-
-    .sliceTileChangeConnected
-      background var(--good)
-
-    .sliceChainId
-      display flex
-      justify-content center
-      align-items center
-      font-family FiraCode, monospace
-      font-size 12px
-      margin 10px
-
-      .sliceChainIdIcon
-        margin-right 6px
-
-  .sliceTile:first-child
-    margin 0px 3px 0px 3px
-
-  .sliceTileHighlight
-    background var(--ghostB)
-
-  .sliceTileClickable:hover
-    background var(--ghostB)
-    box-shadow 0px 2px 2px var(--ghostZ)
-    // animation 4s shake ease-in-out infinite alternate
-
-  .sliceTileClickable:active
-    transform translateY(0px)
-    // animation-name shake-little
-    // animation-duration 2s
-    // animation-timing-function: ease-in-out
-    // animation-iteration-count: infinite
-    // box-shadow 0px 1px 1px var(--ghostZ)
-
-.sliceContainer:last-child
-  margin 0px 6px 0px 6px
-
-
-
-
-
-
-
 .sliceTileGasPrice
   display flex
   justify-content center
@@ -406,7 +217,7 @@
     justify-content flex-end
     align-items center
     box-sizing border-box
-    margin-left 12px
+    margin-left 16px
 
     .gasEstimateRange
       display flex
